+++ mlxtend/classifier/adaline.py
@@ -19,18 +19,19 @@ class Adaline(_BaseClassifier):
     eta : float (default: 0.01)
         solver rate (between 0.0 and 1.0)
     epochs : int (default: 50)
-         Passes over the training dataset.
-    solver : {'gd', 'sgd', 'normal equation'} (default: 'sgd')
-         Method for solving the cost function. 'gd' for gradient descent,
-         'sgd' for stochastic gradient descent, or 'normal equation' (default)
-         to solve the cost function analytically.
-    shuffle : bool (default: False)
-         Shuffles training data every epoch if True to prevent circles.
+        Passes over the training dataset.
+    minibatches : int (default: None)
+        The number of minibatches for gradient-based optimization.
+        If None: Normal Equations (closed-form solution)
+        If 1: Gradient Descent learning
+        If len(y): Stochastic Gradient Descent learning
+        If 1 < minibatches < len(y): Minibatch learning
     random_seed : int (default: None)
         Set random state for shuffling and initializing the weights.
     zero_init_weight : bool (default: False)
         If True, weights are initialized to zero instead of small random
-        numbers in the interval [-0.1, 0.1];
+        numbers following a standard normal distribution with mean=0 and
+        stddev=1;
         ignored if solver='normal equation'
     print_progress : int (default: 0)
         Prints progress in fitting to stderr if not solver='normal equation'
@@ -48,19 +49,15 @@ class Adaline(_BaseClassifier):
       Sum of squared errors after each epoch.
 
     """
-    def __init__(self, eta=0.01, epochs=50, solver='sgd',
-                 random_seed=None, shuffle=False,
+    def __init__(self, eta=0.01, epochs=50,
+                 minibatches=None, random_seed=None,
                  zero_init_weight=False, print_progress=0):
 
         super(Adaline, self).__init__(print_progress=print_progress)
-        np.random.seed(random_seed)
+        self.random_seed = random_seed
         self.eta = eta
+        self.minibatches = minibatches
         self.epochs = epochs
-        self.shuffle = shuffle
-        if solver not in ('normal equation', 'gd', 'sgd'):
-            raise ValueError('learning must be "normal equation", '
-                             '"gd", or "sgd')
-        self.solver = solver
         self.zero_init_weight = zero_init_weight
 
     def fit(self, X, y, init_weights=True):
@@ -96,43 +93,41 @@ class Adaline(_BaseClassifier):
         else:
             self.thres_ = 0.5
 
-        # initialize weights
         if init_weights:
-            self._init_weights(shape=1 + X.shape[1])
+            self.w_ = self._init_weights(shape=1 + X.shape[1],
+                                         zero_init_weight=self.zero_init_weight,
+                                         seed=self.random_seed)
 
         self.cost_ = []
 
-        if self.solver == 'normal equation':
+        if self.minibatches is None:
             self.w_ = self._normal_equation(X, y)
 
         # Gradient descent or stochastic gradient descent learning
         else:
+            n_idx = list(range(y.shape[0]))
             self.init_time_ = time()
             for i in range(self.epochs):
-
-                if self.shuffle:
+                if self.minibatches > 1:
                     X, y = self._shuffle(X, y)
 
-                if self.solver == 'gd':
-                    y_val = self.activation(X)
-                    errors = (y - y_val)
-                    self.w_[1:] += self.eta * X.T.dot(errors)
+                minis = np.array_split(n_idx, self.minibatches)
+                for idx in minis:
+                    y_val = self.activation(X[idx])
+                    errors = (y[idx] - y_val)
+                    self.w_[1:] += self.eta * X[idx].T.dot(errors)
                     self.w_[0] += self.eta * errors.sum()
-                    cost = (errors**2).sum() / 2.0
 
-                elif self.solver == 'sgd':
-                    cost = 0.0
-                    for xi, yi in zip(X, y):
-                        yi_val = self.net_input(xi)
-                        error = (yi - yi_val)
-                        self.w_[1:] += self.eta * xi.dot(error)
-                        self.w_[0] += self.eta * error
-                        cost += error**2 / 2.0
+                cost = self._sum_squared_error_cost(y, self.activation(X))
                 self.cost_.append(cost)
                 if self.print_progress:
                     self._print_progress(epoch=i+1, cost=cost)
 
         return self
+
+    def _sum_squared_error_cost(self, y, y_val):
+        errors = (y - y_val)
+        return (errors**2).sum() / 2.0
 
     def _normal_equation(self, X, y):
         """Solve linear regression analytically."""
@@ -141,14 +136,6 @@ class Adaline(_BaseClassifier):
         z = np.linalg.inv(np.dot(Xb.T, Xb))
         w = np.dot(z, np.dot(Xb.T, y))
         return w
-
-    def _init_weights(self, shape):
-        """Initialize weight coefficients."""
-        if self.zero_init_weight:
-            self.w_ = np.zeros(shape)
-        else:
-            self.w_ = 0.2 * np.random.ranf(shape) - 0.5
-        self.w_.astype('float64')
 
     def net_input(self, X):
         """Compute the linear net input."""