+++ mlxtend/externals/__init__.py
@@ -3,8 +3,3 @@
 # Author: Sebastian Raschka <sebastianraschka.com>
 #
 # License: BSD 3 clause
-
-from .linear_regression import LinearRegression
-from .stacking_regression import StackingRegressor
-
-__all__ = ["LinearRegression", "StackingRegressor"]