+++ mlxtend/feature_selection/sequential_feature_selector.py
@@ -19,6 +19,8 @@ from sklearn.base import clone
 from sklearn.base import BaseEstimator
 from sklearn.base import MetaEstimatorMixin
 from sklearn.cross_validation import cross_val_score
+from ..externals.name_estimators import _name_estimators
+from ..externals import six
 
 
 class SequentialFeatureSelector(BaseEstimator, MetaEstimatorMixin):
@@ -57,7 +59,7 @@ class SequentialFeatureSelector(BaseEstimator, MetaEstimatorMixin):
         algorithm gets stuck in cycles.
     n_jobs : int (default: 1)
         The number of CPUs to use for cross validation. -1 means 'all CPUs'.
-    pre_dispatch : int, or string (default: '2*n_jobs')
+    pre_dispatch : int, or string
         Controls the number of jobs that get dispatched
         during parallel execution in cross_val_score.
         Reducing this number can be useful to avoid an explosion of
@@ -68,7 +70,7 @@ class SequentialFeatureSelector(BaseEstimator, MetaEstimatorMixin):
             to avoid delays due to on-demand spawning of the jobs
         An int, giving the exact number of total jobs that are spawned
         A string, giving an expression as a function
-            of n_jobs, as in `'2*n_jobs'`
+            of n_jobs, as in `2*n_jobs`
 
     Attributes
     ----------
@@ -91,7 +93,7 @@ class SequentialFeatureSelector(BaseEstimator, MetaEstimatorMixin):
                  print_progress=True, scoring='accuracy',
                  cv=5, skip_if_stuck=True, n_jobs=1,
                  pre_dispatch='2*n_jobs'):
-        self.estimator = clone(estimator)
+        self.estimator = estimator
         self.k_features = k_features
         self.forward = forward
         self.floating = floating
@@ -102,8 +104,11 @@ class SequentialFeatureSelector(BaseEstimator, MetaEstimatorMixin):
         self.cv = cv
         self.print_progress = print_progress
         self.n_jobs = n_jobs
+        self.named_est = {key: value for key, value in
+                          _name_estimators([self.estimator])}
 
     def fit(self, X, y):
+        self.est_ = clone(self.estimator)
         if X.shape[1] < self.k_features:
             raise AttributeError('Features in X < k_features')
         if self.skip_if_stuck:
@@ -179,15 +184,15 @@ class SequentialFeatureSelector(BaseEstimator, MetaEstimatorMixin):
 
     def _calc_score(self, X, y, indices):
         if self.cv:
-            scores = cross_val_score(self.estimator,
+            scores = cross_val_score(self.est_,
                                      X[:, indices], y,
                                      cv=self.cv,
                                      scoring=self.scorer,
                                      n_jobs=self.n_jobs,
                                      pre_dispatch=self.pre_dispatch)
         else:
-            self.estimator.fit(X[:, indices], y)
-            scores = np.array([self.scorer(self.estimator, X[:, indices], y)])
+            self.est_.fit(X[:, indices], y)
+            scores = np.array([self.scorer(self.est_, X[:, indices], y)])
         return scores
 
     def _inclusion(self, orig_set, subset, X, y):